+++ ibformers/datasets/ib_extraction/ib_extraction.py
@@ -1,13 +1,15 @@
 import logging
 import os
 from dataclasses import dataclass
-from typing import Tuple, List, Dict, Optional, Sequence, NamedTuple, Iterable
+from typing import Tuple, List, Dict, Optional, Sequence, NamedTuple, Iterable, Any
 
 import datasets
 import numpy as np
 from more_itertools import consecutive_groups
 from typing_extensions import TypedDict
-from ibformers.datasets.ib_common import (
+
+from ibformers.data.utils import ImageProcessor
+from ibformers.datasets.ib_common.ib_common import (
     assert_valid_record,
     Span,
     load_datasets,
@@ -18,12 +20,9 @@ from ibformers.datasets.ib_common import (
     IBDSConfig,
     get_open_fn,
     get_image_features,
+    _DESCRIPTION,
 )
 from instabase.dataset_utils.shared_types import ExtractionFieldDict
-
-_DESCRIPTION = """\
-Internal Instabase Dataset format organized into set of IbDoc files.
-"""
 
 
 # Define some helper classes for easier typing
@@ -173,21 +172,41 @@ class IbExtraction(IbDs):
         data_files = self.config.data_files
         assert isinstance(data_files, dict), "data_files argument should be a dict for this dataset"
         if "train" in data_files:
-            self.datasets_list = load_datasets(data_files["train"], self.config.ibsdk)
-            dataset_classes = self.datasets_list[0].metadata["classes_spec"]["classes"]
-            class_id = self.get_class_id(dataset_classes)
-            schema = dataset_classes[class_id]["schema"]
-            classes = ["O"] + [lab["name"] for lab in schema]
+            features = self.get_train_dataset_features()
+            label2id = features["token_label_ids"].feature._str2int
+            self.config.id2label = {v: k for k, v in label2id.items()}
         elif "test" in data_files:
-            # inference input is a list of parsedibocr files
-            assert self.config.id2label is not None, "Need to pass directly infromation about labels for the inference"
-            classes = [self.config.id2label[i] for i in range(len(self.config.id2label))]
-            if classes[0] != "O":
-                raise logging.error(f"loaded classes does not have required format. No O class: {classes}")
+            features = self.get_inference_dataset_features(self.config)
         else:
             raise ValueError("data_file argument should be either in train or test mode")
 
-        ds_features = get_common_feature_schema(use_image=self.config.use_image)
+        return datasets.DatasetInfo(
+            # This is the description that will appear on the datasets page.
+            description=_DESCRIPTION,
+            features=features,
+            supervised_keys=None,
+        )
+
+    def get_train_dataset_features(self) -> datasets.Features:
+        data_files = self.config.data_files
+        self.datasets_list = load_datasets(data_files["train"], self.config.ibsdk)
+        dataset_classes = self.datasets_list[0].metadata["classes_spec"]["classes"]
+        class_id = self.get_class_id(dataset_classes)
+        schema = dataset_classes[class_id]["schema"]
+        classes = ["O"] + [lab["name"] for lab in schema]
+        return self.create_dataset_features(self.config, classes)
+
+    @classmethod
+    def get_inference_dataset_features(cls, config: IBDSConfig) -> datasets.Features:
+        assert config.id2label is not None, "Need to pass directly infromation about labels for the inference"
+        classes = [config.id2label[i] for i in range(len(config.id2label))]
+        if classes[0] != "O":
+            raise logging.error(f"loaded classes does not have required format. No O class: {classes}")
+        return cls.create_dataset_features(config, classes)
+
+    @staticmethod
+    def create_dataset_features(config, classes):
+        ds_features = get_common_feature_schema(use_image=config.use_image)
         ds_features["token_label_ids"] = datasets.Sequence(datasets.features.ClassLabel(names=classes))
         ds_features["entities"] = datasets.Sequence(
             {
@@ -199,13 +218,7 @@ class IbExtraction(IbDs):
                 "token_label_id": datasets.Value("int64"),
             }
         )
-
-        return datasets.DatasetInfo(
-            # This is the description that will appear on the datasets page.
-            description=_DESCRIPTION,
-            features=datasets.Features(ds_features),
-            supervised_keys=None,
-        )
+        return datasets.Features(ds_features)
 
     def _get_annotation_generator(self, datasets_list: List["DatasetSDK"]) -> Iterable[ExtractionItem]:
         logging.info(f"Reading from Instabase datasets, extracting class name: {self.extraction_class_name}")
@@ -228,10 +241,11 @@ class IbExtraction(IbDs):
                 )
                 yield item
 
-    def _get_annotation_from_model_service(self, records: List[Tuple]) -> Iterable[ExtractionItem]:
+    @classmethod
+    def _get_annotation_from_model_service(self, records: List[Tuple], config: IBDSConfig) -> Iterable[ExtractionItem]:
         # get similar format to the one defined by dataset SDK
         for prediction_item in records:
-            label2ann_label_id = {lab: None for lab in self.config.id2label.values()}
+            label2ann_label_id = {lab: None for lab in config.id2label.values()}
             item = ExtractionItem(
                 ann_item=prediction_item,
                 dataset_id="model-inference",
@@ -240,8 +254,8 @@ class IbExtraction(IbDs):
             )
             yield item
 
-    def process_item(self, item: ExtractionItem) -> Dict:
-        label2id = self.info.features["token_label_ids"].feature._str2int
+    @classmethod
+    def process_item(cls, item: ExtractionItem, config: IBDSConfig, image_processor: ImageProcessor) -> Dict:
         full_path, record_index, record, anno = item.ann_item
 
         anno = {} if anno is None else anno
@@ -261,12 +275,12 @@ class IbExtraction(IbDs):
         doc_id = f"{item.dataset_id}-{os.path.basename(full_path)}-{record_index}.json"
 
         ocr_features = get_ocr_features(words, layouts, doc_id)
-        label_features = get_extraction_labels(ocr_features, anno_fields, label2id, item.label2ann_label_id)
+        label_features = get_extraction_labels(ocr_features, anno_fields, config.label2id, item.label2ann_label_id)
         is_test_file, split = get_extraction_split(anno)
-        open_fn = get_open_fn(self.config.ibsdk)
+        open_fn = get_open_fn(config.ibsdk)
         image_features = (
-            get_image_features(ocr_features, layouts, full_path, open_fn, self.image_processor)
-            if self.config.use_image
+            get_image_features(ocr_features, layouts, full_path, open_fn, image_processor)
+            if image_processor is not None
             else {}
         )